requirements.txt
certifi==2020.6.20
chardet==4.0.0
colorzero==1.1
contourpy==1.0.5
cycler==0.11.0
distro==1.5.0
easygui==0.98.3
fonttools==4.37.2
gpiozero==1.6.2
idna==2.10
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.3
packaging==21.3
picamera2==0.3.3
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
pyparsing==3.0.9
pyserial==3.5
python-apt==2.2.1
python-dateutil==2.8.2
python-prctl==1.7
requests==2.25.1
RPi.GPIO==0.7.0
SciPy==1.8.1
simplejpeg==1.6.4
six==1.16.0
spidev==3.5
ssh-import-id==5.10
toml==0.10.1
urllib3==1.26.5
v4l2-python3==0.3.1
